+++ src/FluxFlow.Components.FileSystem/Contracts/FileWatchChangeType.cs
@@ -1,0 +1,9 @@
+namespace FluxFlow.Components.FileSystem.Contracts;
+
+public enum FileWatchChangeType
+{
+    Created,
+    Changed,
+    Deleted,
+    Renamed
+}