+++ src/FluxFlow.Components.FileSystem/Diagnostics/FileSystemEventNames.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.FileSystem.Diagnostics;
+
+public static class FileSystemEventNames
+{
+    public const string FileWatchChanged = "file.watch.changed";
+}